requirements.txt
aiofiles
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
sqladmin
apscheduler
openai
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
python-dotenv
requests
aiosqlite
httpx==0.23.3
pytest-asyncio>=0.23.0
psycopg2-binary
pre-commit
pytest
